+++ jest.config.js
@@ -16,7 +16,7 @@ const config = {
     "<rootDir>/examples/browser-embedded-sdk-demo-app",
     "<rootDir>/examples/react-sdk-demo-app",
     "<rootDir>/examples/browser-sdk-demo-app",
-    "<rootDir>/packages/indexed-db-stamper",
+    "<rootDir>/packages/indexed-db-stamper/jest.config.js",
   ],
 };