+++ packages/server-sdk/src/index.ts
@@ -2,7 +2,6 @@ import {
   PhantomClient,
   type SignMessageParams,
   type SignAndSendTransactionParams,
-  type SignedTransaction,
   type NetworkId,
   type CreateWalletResult,
   type GetWalletsResult,
@@ -100,7 +99,7 @@ export class ServerSDK  {
     const rawResponse = await this.client.signAndSendTransaction(signAndSendParams);
 
     // Parse the response to get transaction hash and explorer URL
-    return await parseTransactionResponse(rawResponse.rawTransaction, params.networkId);
+    return await parseTransactionResponse(rawResponse.rawTransaction, params.networkId, rawResponse.hash);
   
   }